requirements.txt
Django>=5.2
djangorestframework==3.14.0
whitenoise==6.6.0
django-filter==25.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework-stubs==3.16.7
ruff==0.14.14
psycopg2-binary
celery==5.3.4
redis==5.0.1
django-redis
python-decouple==3.8
drf-spectacular==0.27.1
stripe==8.2.0
django-celery-beat
gunicorn==21.2.0
python-json-logger
django-request-id
django-extensions
django-fsm-2
types-redis
celery-types
python-dotenv
pydantic>=2.0
